canvas/requirements.txt
google-genai
google-cloud-storage
streamlit
st-click-detector